added conversion module for kallistobustools
