+++ conf/modules.config
@@ -80,3 +80,14 @@ if (params.aligner == "star") {
         }
     }
 }
+
+if (params.aligner == "kallisto") {
+    process {
+        withName: MTX_TO_H5AD {
+            publishDir = [
+                path: { "${params.outdir}/kallistobustools/${meta.id}_kallistobustools_count/counts_unfiltered" },
+                mode: params.publish_dir_mode
+            ]
+        }
+    }
+}
